radix_tree: convert to byte-level ART, eliminating Node256 overflow (stage 1)
This is the core of the multi-byte ART refactor.  The tree now operates
internally on raw bytes rather than TokenId values, which eliminates the
"Node256 is full (shouldn't happen in practice)" silent-drop path at
src/radix_tree.hpp:956 by construction: Node256 becomes a true
direct-mapped 256-slot array, with the byte alphabet exactly matching
the slot count.  There is no hashing, no collision handling, and no
overflow case to reason about.

The public API is unchanged — insert/lookup/lookup_instrumented still
take std::span<const TokenId>, and RouteEntry still carries
std::vector<TokenId>.  The conversion to bytes happens in a thin adapter
layer at each entry point via encode_tokens_be() (added in stage 0), and
leaves are decoded back to logical TokenIds when for_each_leaf emits a
RouteEntry.

Key structural changes:
- Node::prefix is now absl::InlinedVector<uint8_t, 32> (same 32-byte
  inline footprint as the previous InlinedVector<TokenId, 8>, just with
  4x the byte capacity).
- Node4/Node16/Node48 keys vectors are now vector<uint8_t>.
- Node256 lost its parallel `keys` array and EMPTY_KEY sentinel — the
  slot index IS the byte key, and an empty slot is simply a null child
  pointer.  1 KB memory savings per Node256.
- key_byte() is deleted; it was the lossy hash that caused the bug.
- find_child / extract_child / set_child / add_child take uint8_t keys
  and have trivial Node256 branches (single array access).  The
  collision-fallback linear scans in Node48 and Node256 are deleted.
- grow_to_node256 is now ~half its former size (no collision handling).
- remove_child and compact handle bytes directly — Node256's remove
  path is a single pointer reset.
- insert_recursive / lookup_recursive / lookup_with_stats operate on
  std::span<const uint8_t> internally; the TokenId public API encodes
  at the entry point.
- for_each_leaf_recursive accumulates bytes during traversal and
  decodes back to TokenIds at leaf emission via decode_tokens_be().  A
  debug assertion verifies each leaf sits at a TokenId-aligned byte
  depth (4-byte multiple from root) — the leaf-alignment invariant
  that the public API guarantees.
- MAX_PREFIX_LENGTH semantics switch from tokens to bytes; bumped from
  256 to 1024 (= 256 tokens * 4 bytes) to preserve the effective
  logical path-compression capacity.

Field renames:
- LookupResult::prefix_tokens_skipped -> prefix_bytes_skipped.  The
  external Prometheus metric `ranvier_radix_tree_average_prefix_skip_length`
  keeps its token-semantic meaning: router_service.cpp:1899 now divides
  by sizeof(TokenId) before recording into the gauge, so dashboards and
  rvctl continue to show tokens-skipped rather than bytes-skipped.  No
  doc breakage, no metric rename.

DumpNode structural change:
- DumpNode::prefix is now vector<uint8_t> and children are
  vector<pair<uint8_t, DumpNode>>.  This is a breaking change to the
  admin HTTP dump format and is intentional: the dump exposes the
  tree's real byte-level structure rather than pretending splits
  respect token boundaries.
- http_controller.cpp's dump_node_to_json casts uint8_t fields to
  unsigned before streaming them — without this cast, uint8_t would
  stream as a character and corrupt the JSON output with non-printable
  bytes.

Tests:
- TokenIdHashCollision_* tests are kept as regression coverage.  The
  comment block above them is rewritten to explain that the scenarios
  they were designed for (tokens sharing the same low byte) are now
  STRUCTURALLY DISTINCT under byte ART, so the tests pass trivially
  rather than exercising collision handling that no longer exists.
- Instrumented tests that checked prefix_tokens_skipped with specific
  expected values are updated to use prefix_bytes_skipped and explain
  the byte-level arithmetic (19 bytes skipped = 5 tokens minus 1 edge
  byte, and 19 / sizeof(TokenId) = 4 matches the original token-level
  expectation).
- New tests added in a "Multi-Byte ART Structural Tests" section:
  * Node256NeverOverflows_256DistinctLowBytes — inserts all 256
    distinct values of a single token's low byte and verifies every
    insert survives; this would have tripped the old silent-drop path.
  * Node256NeverOverflows_AllBytePositions — same, across multiple
    simultaneously-filled Node256s under distinct prefixes.
  * LeafAlignmentInvariant_ForEachLeafYieldsWholeTokens — verifies
    the debug assertion in for_each_leaf_recursive holds across a
    mix of routes that exercise splits, zero-byte paths, and the
    previously-colliding TokenId values.
  * MidTokenSplit_DivergentMidByte — asserts that a split at an
    intra-TokenId byte position (byte 2 of a 4-byte token) still
    leaves both leaves at aligned depths and allows for_each_leaf
    to reassemble tokens correctly.
  * PathCompression_SharedRootPrefixForSmallVocab — verifies the
    big-endian property: small-vocab tokens produce leading zero
    bytes which path compression collapses at the root, yielding
    a non-trivial average prefix length after many inserts.

Docs:
- docs/internals/radix-tree.md updated to describe the byte-level
  Node types, remove references to the old lossy hash and EMPTY_KEY
  sentinel, and update the lookup_instrumented field name.  The
  historical narrative (why the old collision handling existed) is
  preserved as a single paragraph under Node256 for context.

Not touched in this commit (deferred to later stages):
- Benchmark runs to validate performance stays flat or improves.
- Tuning of slab pool sizes for the new node-type distribution.
- Simplification of node_slab_test's internal assertions (the existing
  ones still pass because they only check types and allocation counts,
  neither of which changed).

+++ src/http_controller.cpp
@@ -2585,6 +2585,11 @@ static constexpr int DEFAULT_MAX_DUMP_DEPTH = 32;
 
 // Helper to serialize a DumpNode to JSON with bounded recursion depth.
 // When max_depth is reached, children are replaced with a count summary.
+//
+// Prefix and edge keys are raw bytes after the multi-byte ART refactor.
+// They're emitted as integers (0-255) rather than characters because
+// streaming a uint8_t directly would print it as an ASCII char, which
+// is gibberish for non-printable bytes and would corrupt the JSON.
 static std::string dump_node_to_json(const RadixTree::DumpNode& node, int indent_level = 0, int remaining_depth = DEFAULT_MAX_DUMP_DEPTH) {
     std::ostringstream ss;
     std::string indent(indent_level * 2, ' ');
@@ -2593,11 +2598,11 @@ static std::string dump_node_to_json(const RadixTree::DumpNode& node, int indent
     ss << "{\n";
     ss << inner_indent << "\"type\": \"" << node.type << "\",\n";
 
-    // Prefix array
+    // Prefix array (raw bytes, emitted as integers).
     ss << inner_indent << "\"prefix\": [";
     for (size_t i = 0; i < node.prefix.size(); ++i) {
         if (i > 0) ss << ", ";
-        ss << node.prefix[i];
+        ss << static_cast<unsigned>(node.prefix[i]);
     }
     ss << "],\n";
 
@@ -2611,7 +2616,7 @@ static std::string dump_node_to_json(const RadixTree::DumpNode& node, int indent
     ss << inner_indent << "\"origin\": \"" << node.origin << "\",\n";
     ss << inner_indent << "\"last_accessed_ms\": " << node.last_accessed_ms << ",\n";
 
-    // Children array (bounded by remaining_depth)
+    // Children array (bounded by remaining_depth); edge keys are bytes.
     ss << inner_indent << "\"children\": [";
     if (!node.children.empty()) {
         if (remaining_depth <= 0) {
@@ -2624,7 +2629,7 @@ static std::string dump_node_to_json(const RadixTree::DumpNode& node, int indent
         ss << "\n";
         for (size_t i = 0; i < node.children.size(); ++i) {
             if (i > 0) ss << ",\n";
-            ss << inner_indent << "  {\"edge\": " << node.children[i].first << ", \"node\": ";
+            ss << inner_indent << "  {\"edge\": " << static_cast<unsigned>(node.children[i].first) << ", \"node\": ";
             ss << dump_node_to_json(node.children[i].second, indent_level + 2, remaining_depth - 1);
             ss << "}";
         }
